Plover HID cleanup.
