+++ builddefs/generic_features.mk
@@ -45,6 +45,7 @@ GENERIC_FEATURES = \
     MOUSEKEY \
     MUSIC \
     OS_DETECTION \
+    PLOVER_HID \
     PROGRAMMABLE_BUTTON \
     REPEAT_KEY \
     SECURE \
