+++ quantum/plover_hid.c
@@ -14,10 +14,26 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#pragma once
+#include "plover_hid.h"
+#include "report.h"
+#include "host.h"
 
-#include <stdint.h>
-#include <stdbool.h>
+static report_plover_hid_t plover_hid_report         = {.report_id = REPORT_ID_PLOVER_HID};
+static bool                plover_hid_report_updated = false;
 
-void plover_hid_update(uint8_t button, bool pressed);
-void plover_hid_task(void);
+void plover_hid_update(uint8_t button, bool pressed) {
+    if (pressed) {
+        plover_hid_report.data[button / 8] |= (1 << (7 - (button % 8)));
+    } else {
+        plover_hid_report.data[button / 8] &= ~(1 << (7 - (button % 8)));
+    }
+    plover_hid_report_updated = true;
+}
+
+void plover_hid_task(void) {
+    if (!plover_hid_report_updated) {
+        return;
+    }
+    send_plover_hid(&plover_hid_report);
+    plover_hid_report_updated = false;
+}
