+++ tmk_core/protocol/host.c
@@ -342,6 +342,8 @@ void host_programmable_button_send(uint32_t data) {
 
 __attribute__((weak)) void send_programmable_button(report_programmable_button_t *report) {}
 
+__attribute__((weak)) void send_plover_hid(report_plover_hid_t *report) {}
+
 #ifdef RAW_ENABLE
 void host_raw_hid_send(uint8_t *data, uint8_t length) {
     host_driver_t *driver = host_get_active_driver();
