+++ tmk_core/protocol/host_driver.h
@@ -37,3 +37,4 @@ typedef struct {
 void send_joystick(report_joystick_t *report);
 void send_digitizer(report_digitizer_t *report);
 void send_programmable_button(report_programmable_button_t *report);
+void send_plover_hid(report_plover_hid_t *report);
